requirements.txt
accelerate==0.30.1
deepspeed==0.14.4
einops==0.8.0
transformers==4.43.3
optimum-quanto==0.2.7
datasets==2.3.0
omegaconf==2.3.0
diffusers==0.32.2
sentencepiece==0.2.0
matplotlib==3.10.1
onnxruntime==1.21.0
torchvision==0.21.0
timm==0.6.7
loguru==0.7.3
controlnet_aux==0.0.9
xformers==0.0.29.post3
opencv-python==4.11.0.86
pytorch-lightning==2.5.1
open-clip-torch==2.31.0
torch==2.6.0
came-pytorch==0.1.3
Cython==3.0.12
scipy==1.15.2
safetensors==0.5.3
huggingface_hub==0.25
numpy==2.2.4